Refactor parsing into separate modules
